Circumvent libsvm-3.15 & 3.16 bug (temporary fix for issue #109)

+++ src/machine/cxx/SVM.cc
@@ -196,6 +196,14 @@ boost::shared_ptr<svm_model> mach::svm_unpickle
   boost::shared_ptr<svm_model> retval(svm_load_model(tmp_filename), 
       std::ptr_fun(svm_model_free));
 
+  if (!retval) {
+    boost::format s("cannot open model file '%s'");
+    s % tmp_filename;
+    throw std::runtime_error(s.str());
+  }
+
+  retval->sv_indices = 0; ///< force initialization: see ticket #109
+
   //unlinks the temporary file
   boost::filesystem::remove(tmp_filename); 
 
@@ -231,6 +239,7 @@ mach::SupportVector::SupportVector(const std::string& model_file):
     s % model_file;
     throw std::runtime_error(s.str());
   }
+  m_model->sv_indices = 0; ///< force initialization: see ticket #109
   reset();
 }
